Handle images more like links.
Remove MD_SPAN_IMG_DETAIL::alt. Instead, the contents of the image is
propagated to the renderer via MD_RENDERER::text() callback.

 * This fixes handling of entities inside the image text (issue #4).
 * It simplifies parsing and, more importantly, it better distingusshes
   what is responsibility of parser or renderer respectively.
 * This allows more flexibility on renderers side. Renderer who do not
 * really support images can just output the image content as any
   other text.

The cost is a renderer into HTML (if it wants to render image contents
into the attribute ALT of the IMG tag), has to handle images with more
care. Typically such renderer has to track whether it is inside an image,
and if so, then render span enter/leave as an empty string.