Fix conflict in parsing permissive autolinks and ordinary links.
The issues is caused by the fact that we do not know exact position
of permissive auto-link in time of md_collect_marks() because there
is no syntax to mark its end on the 1st place.

This causes that eventually, the closer mark in ctx->marks[] can be
out-of-order somewhat.

As a consequence, if some other mark range (e.g. ordinary link)
shadows the auto-link, the closer mark may be left outside the shadowed
range and survive till the phase when we generate the output.

We fix by using an extra mark flag to remember we did really output
the opener mark, and output the closer only in such case.

Fixes #53.