md_is_code_span: Optimize scanning for the closer mark.
When scanning for the closer mark, we remember (the latest) position of
potential closers for all other lengths as well.

This means that:

(1) If we find it, we reduced the task because all subsequent scan shall
begin after the closer.

(2) If we do not find it, then we have to reach the end of the block and
hence we then know (for every allowed marker length) the position of last
such backtick sequence.

(3) That makes the guaranty that any subsequent call with either succeed
in its scan (and reduce the task even further); or that we shall be able
to detect instantly there is no suitable closer.

I.e. every call either reduces the task by O(n) scan (1); or collects
all the data in O(n) because (2) happens at most once; or fails in O(1)
(3).

This makes O(n) guaranty of the function complexity.