Simplify parsing of tables (#97)
We do so by removing the function md_is_table_row().

md_is_table_row() did some crazy inline parsing to detect whether the
line contains at least one pipe which is not inside a code span or other
high-priority inline element.

This was very complicated under the hood and to was actually breaking
the clean design which separates block analysis parse and inline analysis
of each block contents.

We now just use the table underline for determining the block is table
and its properties like e.g. the column count.

This means a paragraph now cannot interrupt a table. This is a change in
a behavior but likely acceptable one as it actually brings the behavior
closer to behavior of tables in cmark-gfm in this regard.

Last but not least, it seems to prevent adoption of other useful
features, for about that, see the discussion in PR #92.