Introduce MD_PARSER_v2.
This version of MD_PARSER changes prototypes of callbacks enter_block(),
leave_block(), enter_span(), leave_span() and text() so that their 1st
argument is int instead of enumerations MD_BLOCK_TYPE, MD_SPAN_TYPE and
MD_TEXT_TYPE respectivelly.

The purpose if this is to enable introduction of new block, span and
text type in runtime in upcoming patches.